backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.23
pydantic==1.10.15
python-dotenv==1.0.1
alembic==1.13.1
python-multipart==0.0.9